Added redirect to "auth required" page if no authentication
This is not likely to happen, as CAS should not send the user back
unless they're authenticated and the plugin should redirect to CAS if
the user is not already authenticated. This exists on the off-chance
that something happens to cause the user to view the page w/o being
forced to log in.